requirements.txt
selenium==4.15.0 #automatización web
pytest==7.4.3 #testing
pytest-xdist==3.3.1 
pytest-html==4.1.1 
pytest-metadata==3.0.0
allure-pytest==2.13.2 #reportes profesionales
allure-python-commons==2.13.2
opencv-python==4.8.1.78 #visión por computadora
pyautogui==0.9.54 #automatización de escritorio
Pillow==10.3.0 #procesamiento de imágenes
faker==19.6.2 #datos de prueba
#sqlite3
python-dateutil==2.8.2
webdriver-manager==4.0.1